+++ apps/client-dev/component/MarkdownPage.tsx
@@ -1,7 +1,10 @@
-import type { VNode } from "@elf-framework/sapa";
+import { Divider } from "@elf-framework/ui";
 
 import "./MarkdownPage.scss";
-import { TableOfContents, TableOfContentsItem } from "./TableOfContents";
+import { TableOfContents } from "./TableOfContents";
+import { makeId } from "./utils/url";
+
+import { FrontMatter } from "~/types/site";
 
 function FileEditorLink({ filename }: { filename: string }) {
   return filename ? (
@@ -17,11 +20,65 @@ function FileEditorLink({ filename }: { filename: string }) {
   );
 }
 
-interface MarkdownPageProps {
+interface MarkdownPageProps extends FrontMatter {
   page: any;
   filename?: string;
   tableOfContents?: boolean;
   editableSourceLink?: boolean;
+}
+
+function Heading1(props: any) {
+  const id = ["1", makeId(props.content.join(""))].join("-");
+  return (
+    <h1 id={id} class="h1">
+      {props.content}
+    </h1>
+  );
+}
+
+function Heading2(props: any) {
+  const id = ["2", makeId(props.content.join(""))].join("-");
+  return (
+    <h2 id={id} class="h2">
+      {props.content}
+    </h2>
+  );
+}
+
+function Heading3(props: any) {
+  const id = ["3", makeId(props.content.join(""))].join("-");
+  return (
+    <h3 id={id} class="h3">
+      {props.content}
+    </h3>
+  );
+}
+
+function Heading4(props: any) {
+  const id = ["4", makeId(props.content.join(""))].join("-");
+  return (
+    <h4 id={id} class="h4">
+      {props.content}
+    </h4>
+  );
+}
+
+function Heading5(props: any) {
+  const id = ["5", makeId(props.content.join(""))].join("-");
+  return (
+    <h5 id={id} class="h5">
+      {props.content}
+    </h5>
+  );
+}
+
+function Heading6(props: any) {
+  const id = ["6", makeId(props.content.join(""))].join("-");
+  return (
+    <h6 id={id} class="h6">
+      {props.content}
+    </h6>
+  );
 }
 
 export function MarkdownPage({
@@ -31,24 +88,22 @@ export function MarkdownPage({
   tableOfContents = false,
   // menuLink = false,
   editableSourceLink = false,
+
+  /** links in markdown paper */
+  displayLinks = false,
+  links = [],
 }: MarkdownPageProps) {
-  const template = Page();
+  // fixed table of contents
 
-  const items: TableOfContentsItem[] = [];
-  template.children.forEach((child: VNode) => {
-    if (child.nodeName?.startsWith("H")) {
-      const text = child.makeText(" ");
-      const id = child.makeText("-");
-
-      // element 에 적용이 되기 위해서 memoizedProps 를 변경
-      const targetId = encodeURIComponent(id);
-      child.props.id = targetId;
-      child.memoizedProps.id = targetId;
-
-      const level = child.nodeName.replace("H", "");
-
-      items.push({ id, text, level });
-    }
+  const template = Page({
+    components: {
+      h1: Heading1,
+      h2: Heading2,
+      h3: Heading3,
+      h4: Heading4,
+      h5: Heading5,
+      h6: Heading6,
+    },
   });
 
   return (
@@ -60,6 +115,22 @@ export function MarkdownPage({
           ) : undefined}
           <div class="content-inner" ref="$inner">
             {template.children || template}
+
+            {displayLinks ? <Divider /> : undefined}
+
+            {displayLinks ? (
+              <div>
+                <ul>
+                  {links.map((link) => (
+                    <li>
+                      <a href={link.url} target="_link">
+                        {link.title}
+                      </a>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            ) : undefined}
           </div>
           {/* <Divider margin={100} /> */}
           {/* {menuLink ? <LinkedPage menu={menu} /> : undefined} */}
@@ -69,9 +140,7 @@ export function MarkdownPage({
           ) : undefined}
         </div>
 
-        {items.length && tableOfContents ? (
-          <TableOfContents items={items} />
-        ) : undefined}
+        {tableOfContents ? <TableOfContents items={template} /> : undefined}
       </div>
     </div>
   );