+++ apps/client-dev/component/site/Body.tsx
@@ -1,6 +1,10 @@
+import type { ContentType } from "@elf-framework/sapa";
+
 import "./Body.scss";
 
-interface BodyProps {}
+interface BodyProps {
+  content: ContentType;
+}
 
 export function Body(props: BodyProps) {
   const { content } = props;