+++ apps/client-dev/pages/article/index.tsx
@@ -2,12 +2,18 @@ import { start } from "@elf-framework/sapa";
 
 // real view page
 import Page, * as props from "./index.mdx";
+import metaJSON from "./index.meta.json";
 
 import { Site } from "~/component/site/Site";
 
+const newProps = {
+  ...metaJSON,
+  ...props,
+};
+
 const filename = "pages/article/index.mdx";
 
-const site = <Site filename={filename} page={Page} {...props} />;
+const site = <Site filename={filename} page={Page} {...newProps} />;
 
 start(site, {
   container: document.getElementById("app"),